Change `Rgb` values to floats and remove range limits
Change all `Rgb` color components to floats and remove the restriction that they have to be within
the 0-255 range after every operation. This maintains accuracy during calculations until the
components are forced into this range with the new `clamp` method. This is a widespread change due
to all code using the `Rgb` struct needing to be changed to use f64s rather than u8s.