Fix clipping bugs with triangle winding and normals
The code to clip a triangle against a clipping plane has been overhauled.
Previously, the order of vertexes in the original triangle was not preserved
in replacement triangles, leading to some facing the wrong way and
therefore getting culled. The new code is simpler, but has several similar
blocks that could be abstracted but only if vertex order is preserved.